Simple CRUD done.

+++ demo/src/main/java/com/example/demo/repository/normal_repos.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entities.normal_entity;
 
-public interface normal_repos extends JpaRepository<normal_entity, Integer> {
+public interface normal_repos extends JpaRepository<normal_entity, Long> {
 
 }
